refactor(rust): establish migration workspace foundation
