+++ .github/simplicity-budgets.toml
@@ -45,6 +45,8 @@ allowed = [
   "AGENTS.md",
   "CHANGELOG.md",
   "CLAUDE.md",
+  "Cargo.lock",
+  "Cargo.toml",
   "Dockerfile",
   "Dockerfile.distroless",
   "LICENSE",
@@ -54,7 +56,9 @@ allowed = [
   "README.zh-CN.md",
   "app",
   "config",
+  "crates",
   "deploy",
+  "deny.toml",
   "docker-compose.prod.yml",
   "docker-compose.yml",
   "docs",
@@ -65,6 +69,7 @@ allowed = [
   "openspec",
   "pyproject.toml",
   "renovate.json",
+  "rust-toolchain.toml",
   "scripts",
   "tests",
   "uv.lock",
