+++ .gitignore
@@ -22,7 +22,8 @@ __pycache__/
 build/
 dist/
 site/
-native/**/target/
+/target/
+crates/**/target/
 app/static/
 deploy/helm/*/charts/
