+++ Cargo.toml
@@ -1,10 +1,20 @@
-[package]
-name = "codex-lb-native-egress"
+[workspace]
+resolver = "3"
+members = [
+    "crates/codex-lb-egress",
+    "crates/codex-lb-egress-worker",
+    "crates/codex-lb-protocol",
+]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2024"
+rust-version = "1.96"
+license = "MIT"
+repository = "https://github.com/Soju06/codex-lb"
 publish = false
 
-[dependencies]
+[workspace.dependencies]
 aws-lc-rs = "=1.16.2"
 base64 = "0.22.1"
 futures-util = { version = "0.3.31", features = ["sink"] }
@@ -21,6 +31,19 @@ tokio-rustls = { version = "=0.26.4", default-features = false, features = ["aws
 tokio-tungstenite = { version = "0.28.0", features = ["proxy", "rustls-tls-native-roots"] }
 tungstenite = { version = "0.27.0", features = ["deflate", "proxy"] }
 url = "2.5.7"
+
+[workspace.lints.rust]
+unsafe_code = "forbid"
+
+[workspace.lints.clippy]
+dbg_macro = "deny"
+todo = "deny"
+unimplemented = "deny"
+
+[profile.release]
+lto = "thin"
+codegen-units = 1
+strip = "symbols"
 
 [patch.crates-io]
 tokio-tungstenite = { git = "https://github.com/openai-oss-forks/tokio-tungstenite", rev = "0e5b2d73aa18dd9f0a50ee9ff199d5aef7594186" }
