+++ Makefile
@@ -56,6 +56,8 @@ help:
 	  '  make lint                    ruff check + format check + architecture checks' \
 	  '  make architecture-check      proxy architecture fitness ratchets' \
 	  '  make typecheck               ty check' \
+	  '  make rust-check              fmt + clippy + tests + release build' \
+	  '  make rust-audit              cargo-deny dependency policy' \
 	  '  make frontend-test           vitest coverage, same as CI' \
 	  '  make test-dashboard-browser-smoke  built dashboard against the real local API' \
 	  '  make test-unit               unit pytest slice, same as CI' \
@@ -91,7 +93,7 @@ test-dashboard-browser-smoke: frontend-build frontend-playwright-chromium
 	uv sync --dev --frozen
 	uv run python scripts/run_dashboard_browser_smoke.py --frontend-built
 
-.PHONY: lint typecheck architecture-check
+.PHONY: lint typecheck architecture-check rust-fmt rust-lint rust-test rust-build rust-check rust-audit
 lint: architecture-check
 	uv run ruff check .
 	uv run ruff format --check .
@@ -102,6 +104,23 @@ architecture-check:
 typecheck:
 	uv sync --dev --frozen
 	uv run ty check
+
+rust-fmt:
+	cargo fmt --all -- --check
+
+rust-lint:
+	cargo clippy --workspace --all-targets --all-features --locked -- -D warnings
+
+rust-test:
+	cargo test --workspace --all-targets --locked
+
+rust-build:
+	cargo build --release --locked --package codex-lb-egress-worker --bin codex-lb-native-egress
+
+rust-check: rust-fmt rust-lint rust-test rust-build
+
+rust-audit:
+	cargo deny --all-features check
 
 .PHONY: test-unit test-integration-core test-integration-core-shard \
 	test-integration-core-1 test-integration-core-2 test-integration-core-3 \
@@ -228,8 +247,8 @@ helm-smoke-kind:
 	KUBE_CONTEXT=kind-codex-lb-smoke IMAGE_REGISTRY=ghcr.io IMAGE_REPOSITORY=soju06/codex-lb IMAGE_TAG=ci ./scripts/helm-kind-smoke.sh external-db
 
 .PHONY: ci-fast ci
-ci-fast: lint typecheck frontend-test test-unit package
+ci-fast: lint typecheck rust-check frontend-test test-unit package
 
-ci: frontend-lint frontend-typecheck frontend-test frontend-build lint typecheck \
+ci: frontend-lint frontend-typecheck frontend-test frontend-build lint typecheck rust-check \
 	test-unit test-integration-core test-integration-bridge test-e2e test-postgres \
 	migration-check migration-check-postgres package docker helm-check helm-smoke-kind
