+++ crates/codex-lb-egress/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "codex-lb-egress"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+aws-lc-rs.workspace = true
+base64.workspace = true
+codex-lb-protocol = { version = "0.1.0", path = "../codex-lb-protocol" }
+futures-util.workspace = true
+hyper.workspace = true
+hyper-rustls.workspace = true
+hyper-util.workspace = true
+reqwest.workspace = true
+rustls.workspace = true
+rustls-native-certs.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tokio-rustls.workspace = true
+tokio-tungstenite.workspace = true
+tungstenite.workspace = true
+url.workspace = true
