+++ crates/codex-lb-protocol/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "codex-lb-protocol"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
