+++ mkdocs.yml
@@ -73,4 +73,6 @@ nav:
   - Reference:
       - Settings: reference/settings.md
   - Traffic Parity: traffic-parity.md
+  - Development:
+      - Rust Migration Architecture: rust-architecture.md
   - Troubleshooting: troubleshooting.md
