+++ openspec/changes/archive/2026-08-31-establish-rust-workspace-foundation/tasks.md
@@ -1,0 +1,17 @@
+## 1. Workspace and boundaries
+
+- [x] 1.1 Create the root virtual workspace, lockfile, toolchain, and policy.
+- [x] 1.2 Separate protocol, reusable egress, and worker binary crates.
+- [x] 1.3 Move container builds to the root locked workspace.
+
+## 2. Compatibility
+
+- [x] 2.1 Add version/capability hello messages and fail-closed Rust gating.
+- [x] 2.2 Add bounded Python negotiation before generation assignment.
+- [x] 2.3 Add a shared cross-language fixture and incompatibility coverage.
+
+## 3. Governance and verification
+
+- [x] 3.1 Add formatting, Clippy, workspace tests, release build, and cargo-deny CI.
+- [x] 3.2 Document final-state crate boundaries and slice migration rules.
+- [x] 3.3 Run Rust, Python adapter/packaging, OpenSpec, and container-path validation.
