+++ openspec/changes/archive/2026-08-31-ratchet-native-parity-proxy-architecture/tasks.md
@@ -1,0 +1,10 @@
+## 1. Architecture policy
+
+- [x] 1.1 Measure post-cutover façade and streaming-mixin sizes.
+- [x] 1.2 Reset the ratchets to those exact values.
+- [x] 1.3 Declare the HTTP-bridge to streaming-policy dependency explicitly.
+
+## 2. Verification
+
+- [x] 2.1 Run the repository architecture checker and its unit coverage.
+- [x] 2.2 Validate the updated proxy-architecture specification strictly.
